pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-lifecycle-platform"
version = "0.1.0"
description = "Prompt lifecycle and experimentation platform with registry, evaluation, regression checks, and A/B simulation."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "OpenAI Codex" }]
dependencies = [
  "pydantic>=2.7,<3",
  "PyYAML>=6.0",
  "typer>=0.12",
  "rich>=13.7",
  "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5",
]

[project.scripts]
prompt-platform = "prompt_platform.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py39"
